requirements.txt
Flask==3.1.0
PyJWT==2.8.0
firebase-admin==6.2.0
gunicorn==20.1.0
numpy==1.24.2
tensorflow>=2.18.0
pillow==10.0.0
google-cloud-storage==2.9.0
google-cloud-firestore==2.10.0
protobuf>=3.20.3,<5.0.0dev
pytz==2023.3
